refactor(server): split eviction at its two passes to stay inside the line limit
The mutation fix in 8b972de took eviction.rs from 499 lines to 569, across
CONTRIBUTING's 500-line rule, and scripts/check_file_lengths.sh caught it. That
is the check working: the file was one line under before, so any addition at all
would have crossed it.

Recording an exemption was the sanctioned alternative and would have been the
wrong one. The file already documented the seam it needed — "two passes, on two
schedules, because they cost different things" — so the split was waiting to be
made rather than invented to satisfy a counter:

  eviction/mod.rs       the schedule. EvictionPasses, should_evict, maybe_evict,
                        evict, and the amortized O(n) TTL pass.
  eviction/capacity.rs  the pass that runs on every write once the store is
                        full. EVICTION_SCAN_WINDOW, evict_over_capacity,
                        capacity_victims, is_terminal. Its bound is the whole
                        subject of the file, which is the point of separating it.
  eviction/fixtures.rs  store_of/config/ids. Shared rather than copied because
                        both suites must agree on what "oldest first" means; two
                        copies that drifted would let one stop matching the other
                        without either failing.

The TTL and capacity bodies are now named functions instead of inline branches
of evict, leaving evict as the gate it actually is.

No behaviour change. The same 11 tests pass, unmodified apart from their module
homes. check_file_lengths.sh returns to "79 of 387 tracked source files exceed
500 lines, all recorded" — the ratchet list did not grow — and
check_mutation_scope.sh still reports 141 of 141, so all three files remain
inside the per-PR mutation gate rather than slipping out of it.

Also: the local fmt check that let 8b972de through was unsound. It read
`cargo fmt --all --check | head -20 && echo "FMT OK"`, where && binds to head's
exit status, so it printed OK regardless — and the four non-rustfmt steps of the
Format job were never run locally at all. All five are checked individually here.
